Packages renamed
eu.ortlepp.blogbuilder.tools renamed to eu.ortlepp.blogbuilder.util
