+++ src/main/java/eu/ortlepp/blogbuilder/action/Initialize.java
@@ -1,13 +1,13 @@
 package eu.ortlepp.blogbuilder.action;
 
-import eu.ortlepp.blogbuilder.tools.config.Directories;
-
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.logging.Logger;
+
+import eu.ortlepp.blogbuilder.util.config.Directories;
 
 /**
  * Action: Initialize a new project.
