+++ src/main/java/eu/ortlepp/blogbuilder/util/AbstractXmlCreator.java
@@ -1,4 +1,4 @@
-package eu.ortlepp.blogbuilder.tools;
+package eu.ortlepp.blogbuilder.util;
 
 import org.w3c.dom.Attr;
 import org.w3c.dom.Document;
