+++ src/main/java/eu/ortlepp/blogbuilder/util/ResourceCopy.java
@@ -1,6 +1,4 @@
-package eu.ortlepp.blogbuilder.tools;
-
-import eu.ortlepp.blogbuilder.tools.config.Directories;
+package eu.ortlepp.blogbuilder.util;
 
 import java.io.IOException;
 import java.nio.file.FileVisitResult;
@@ -10,6 +8,8 @@ import java.nio.file.Paths;
 import java.nio.file.SimpleFileVisitor;
 import java.nio.file.attribute.BasicFileAttributes;
 import java.util.logging.Logger;
+
+import eu.ortlepp.blogbuilder.util.config.Directories;
 
 /**
  * A tool class to copy the contents from the resources directory to the directory with the built blog.
