+++ src/main/java/eu/ortlepp/blogbuilder/util/Scanner.java
@@ -1,4 +1,4 @@
-package eu.ortlepp.blogbuilder.tools;
+package eu.ortlepp.blogbuilder.util;
 
 import eu.ortlepp.blogbuilder.model.Document;
